refactor(course-promotion): add try-catch + add missing awaits
