+++ common/mails/courses/index.ts
@@ -17,6 +17,7 @@ import { getCourseCapacity, isParticipant } from '../../courses/participants';
 import { getCourseImageURL } from '../../courses/util';
 import { createSecretEmailToken } from '../../secret';
 import { getCourse } from '../../../graphql/util';
+import { shuffleArray } from '../../../common/util/basic';
 
 const logger = getLogger('Course Notification');
 
@@ -234,19 +235,8 @@ const isPromotionValid = (publishedAt: Date, capacity: number, alreadyPromoted: 
     return capacity < 0.75 && alreadyPromoted === false && (daysDiff === null || daysDiff > 3);
 };
 
-function shuffleArray<T>(array: T[]): T[] {
-    const shuffledArray = [...array]; // Create a copy of the original array
-
-    for (let i = shuffledArray.length - 1; i > 0; i--) {
-        const j = Math.floor(Math.random() * (i + 1));
-        [shuffledArray[i], shuffledArray[j]] = [shuffledArray[j], shuffledArray[i]];
-    }
-
-    return shuffledArray;
-}
-
 export async function sendPupilCoursePromotion(subcourse: Prisma.subcourse) {
-    const predictedPupilResponseRate = 10; // in percent
+    const predictedPupilResponseRate = 5; // in percent
 
     const courseCapacity = await getCourseCapacity(subcourse);
     const { alreadyPromoted, publishedAt } = subcourse;
@@ -267,9 +257,9 @@ export async function sendPupilCoursePromotion(subcourse: Prisma.subcourse) {
     });
 
     const courseSubject = course.subject;
-    const filteredPupils = pupils.filter((pupil) => {
+    const filteredPupils = pupils.filter(async (pupil) => {
         const subjects = parseSubjectString(pupil.subjects);
-        const pupilAlreadyParticipant = isParticipant(subcourse, pupil);
+        const pupilAlreadyParticipant = await isParticipant(subcourse, pupil);
         const isPupilsSubject = subjects.some((subject) => subject.name == courseSubject);
         return !courseSubject || (!pupilAlreadyParticipant && isPupilsSubject);
     });
@@ -287,7 +277,7 @@ export async function sendPupilCoursePromotion(subcourse: Prisma.subcourse) {
     );
 
     const context = await getNotificationContextForSubcourse(course, subcourse);
-    Notification.bulkActionTaken(
+    await Notification.bulkActionTaken(
         randomFilteredPupilSample.map((pupil) => userForPupil(pupil)),
         'available_places_on_subcourse',
         context
