+++ common/util/basic.ts
@@ -75,3 +75,14 @@ export function isEmail(email: string): boolean {
         email
     );
 }
+
+export function shuffleArray<T>(array: T[]): T[] {
+    const shuffledArray = [...array]; // Create a copy of the original array
+
+    for (let i = shuffledArray.length - 1; i > 0; i--) {
+        const j = Math.floor(Math.random() * (i + 1));
+        [shuffledArray[i], shuffledArray[j]] = [shuffledArray[j], shuffledArray[i]];
+    }
+
+    return shuffledArray;
+}
